+++ libzfs-sys/build.rs
@@ -16,7 +16,10 @@ fn main() {
 
     env::set_var("LIBCLANG_PATH", "/opt/llvm-5.0.0/lib64/");
 
-    pkg_config::Config::new().probe("libzfs").unwrap();
+    pkg_config::Config::new()
+        .atleast_version("0.7.9")
+        .probe("libzfs")
+        .unwrap();
     println!("cargo:rustc-link-lib=zpool");
 
     // Skip building if bindings already exist.
