+++ libzfs-types/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "libzfs-types"
+version = "0.1.0"
+authors = ["IML Team <iml@whamcloud.com>"]
+description = "Shared types for libzfs"
+license = "MIT"
+
+[dependencies]
+serde = "1.0"
+serde_derive = "1.0"
