+++ libzfs/Cargo.toml
@@ -6,7 +6,8 @@ description = "Rust wrapper around libzfs-sys"
 license = "MIT"
 
 [dependencies]
-libzfs-sys = { path = "../libzfs-sys", version = "0.5.7" }
+libzfs-sys = { path = "../libzfs-sys", version = "0.5.7"}
+libzfs-types = { path = "../libzfs-types", version = "0.1.0" }
 nvpair-sys = "0.1"
 serde = "1.0"
 serde_derive = "1.0"
